video-monitoring-handson/sample01/backend/requirements.txt
asyncio==3.4.3
fastapi==0.116.1
gunicorn==23.0.0
google-auth==2.40.3
google-genai==1.25.0
uvicorn==0.35.0
python-dotenv==1.1.1
